Move ke_c_sw to c_sw (#178)
