+++ tests/translate/translate_c_sw.py
@@ -154,3 +154,52 @@ class TranslateCirculation_Cgrid(TranslateFortranData2Py):
             domain=self.grid.domain_shape_compute_buffer_2d(add=(1, 1, 0)),
         )
         return self.slice_output({"vort_c": inputs["vort_c"]})
+
+
+class TranslateKE_C_SW(TranslateFortranData2Py):
+    def __init__(self, grid):
+        super().__init__(grid)
+        self.in_vars["data_vars"] = {
+            "uc": {},
+            "vc": {},
+            "u": {},
+            "v": {},
+            "ua": {},
+            "va": {},
+        }
+        self.in_vars["parameters"] = ["dt2"]
+        self.out_vars = {
+            "ke_c": {
+                "istart": grid.is_ - 1,
+                "iend": grid.ie + 1,
+                "jstart": grid.js - 1,
+                "jend": grid.je + 1,
+            },
+            "vort_c": {
+                "istart": grid.is_ - 1,
+                "iend": grid.ie + 1,
+                "jstart": grid.js - 1,
+                "jend": grid.je + 1,
+            },
+        }
+
+    def compute(self, inputs):
+        self.make_storage_data_input_vars(inputs)
+        ke = utils.make_storage_from_shape(inputs["uc"].shape)
+        vort = utils.make_storage_from_shape(inputs["vc"].shape)
+        c_sw.update_vorticity_and_kinetic_energy(
+            ke=ke,
+            vort=vort,
+            sin_sg1=self.grid.sin_sg1,
+            cos_sg1=self.grid.cos_sg1,
+            sin_sg2=self.grid.sin_sg2,
+            cos_sg2=self.grid.cos_sg2,
+            sin_sg3=self.grid.sin_sg3,
+            cos_sg3=self.grid.cos_sg3,
+            sin_sg4=self.grid.sin_sg4,
+            cos_sg4=self.grid.cos_sg4,
+            **inputs,
+            origin=(self.grid.is_ - 1, self.grid.js - 1, 0),
+            domain=(self.grid.nic + 2, self.grid.njc + 2, self.grid.npz),
+        )
+        return self.slice_output(inputs, {"ke_c": ke, "vort_c": vort})
